+++ packages/react-core/src/components/Nav/examples/NavDrilldown.tsx
@@ -65,7 +65,7 @@ export const NavLevelThreeDrill: React.FunctionComponent = () => {
   };
 
   return (
-    <Nav>
+    <Nav aria-label="Drilldown global nav">
       <Menu
         id="rootMenu"
         containsDrilldown