+++ packages/react-tokens/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 4.77.3 (2022-07-15)
+
+**Note:** Version bump only for package @patternfly/react-tokens
+
+
+
+
+
+## 4.77.2 (2022-07-15)
+
+**Note:** Version bump only for package @patternfly/react-tokens
+
+
+
+
+
+## 4.77.1 (2022-07-15)
+
+**Note:** Version bump only for package @patternfly/react-tokens
+
+
+
+
+
 # 4.77.0 (2022-07-14)